Split core/scrollbar.js into 2 files
